[ci] Revert "firtool" artifact name change
The name of the artifact that include the nightly and published `firtool`
was changed from `firrtl-bin` to `firtool-bin`.  This caused a bunch of
other CI to stop getting the latest version (and will break CI on the
1.52.0 release).  I'm fine with the name change.  I would like to delay
the name change until a later point.  This reverts the name change back to
the original.

CC: @SpriteOvO

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>