[ESI] Introduce new type of cosim: low level
The new "low level" cosim emulates a PCIe device. It will provide a few
raw interfaces which are quite similar to a PCIe connected board: MMIO
space, host memory access, and a configurable number of DRAM channels.

This type of cosim allows for more code sharing with real systems. So
there'll be less code to write and it'll be easier to test more of the
stuff we do write.

Whether or not this is a permanent replacement for the existing cosim is
yet to be seen. That the existing cosim is so different from a real
system is a constant source of pain. The existing one is faster, though,
since it directly implements channel connectivity. Whether or not that
speed is worth the additional code and maintance burden is yet to be
seen.