[FIRRTL][LowerToHW] Copy InnerSymAttrs to lowered ops
This change ensures that when lowering InnerSymbol operations, we copy the
existing inner symbol to the target, as well as properly adding a symbol on
fieldID 0 when required.  This fixes an assumption of symbol name uniqueness,
and they are now uniqued in the module's symbol namespace. A side effect of
this change is that the used symbol names are always of the form `@sym_1`.
This also fixes a bug where an InnerSymAttr was assumed to target field ID 0,
and could result in an inccorect lowering of the DontTouch annotation.