[FIRRTL] Prefix for Seq Mem Group Extract (#5915)
Fix a bug in the composition of the PrefixModules and ExtractInstances
passes.  Because the latter can generate modules with new names when
using the grouping feature, this module needs a prefix and does not
currently get one.  Change the PrefixModules pass to add a wildcat
attribute "prefix" to the DUT annotation.  This prefix is then used by
ExtractInstances to properly generate the module name.

This is a variant of the solution used for the GrandCentral pass where
the companion module annotation is given a wildcat attribute that stores
the prefix which is then used by the pass.

A better solution would be to reorder PrefixModules after
ExtractInstances.  However, this has complications with verbatim
metadata emission and the possibility of PrefixModules manipulating the
hierarchy (duplicating modules to resolve non-local prefixes).  There is
more discussion about this on the #4593.

Fixes #4593.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>
Co-authored-by: Jack Koenig <koenig@sifive.com>