[FIRRTL] re-work PathOp to use annotations (#5933)
The PathOp is being re-worked to remotely reference hardware objects using
(possibly non-local) annotations.  The hardware target recieves an annotation
with class `circt.tracker` and an identifier.  The Path operation tracks the
original reference kind and the target identifier. It is legal to optimize away
the hardware target, which will leave the Path op with a dangling identifier
reference. A later pass will lower this pick bb5ad82a6 [FIRRTL] Add enumeration
for OM path kinds FIRRTL PathOp to an OM specific Path operation, which will
not use annotations.  See the example in FIRRTLExpressions.td for what this
looks like in FIRRTL.