[OM] Class ops should not be symbol tables
Currently, OM class operations are symbols tables and their fields are symbols,
which is used to facilitate the lookup of fields.  The ObjectFieldOp verifier
needs to verify that it is accessing a legal field of the class, and does this
by first looking up the class in the MLIR ModuleOp, and then the field in the
OM ClassOp.  This double lookup strategy is problematic because it does not
follow the normal rules of symbol resolution as defined by the SymbolTable.
This means that anything that examines symbol uses (e.g. SymbolDCE) will be
incorrect.  We have faced this kind of problem before, which is why we invented
InnerSymbolTables and InnerSymbols.

This is also needed to properly support path operations: path ops will have a
references to HierPathOps and we need lookup to work correctly. For this, we
will need two things:
1. Since HierPathOps will live inside a FIRRTL circuit, and FIRRTL circuits are
   themselves symbol tables, we will need to put OM classes under the circuit
   as well.
2. When this happens, it will no longer be acceptable for the ObjectFieldOp
   verifier to "jump up" to the MLIRModule op's symbol table for lookup, since
   the classes will actually be under the FIRRTL circuit op.

By removing the SymbolTable interface from class operations, we can change
symbol resolution to use the normal rules, which enables both path operations
to reference NLAs and OM class operations to live in FIRRTL circuits.

This PR removes the SymbolTable op interface from ClassOps, and switches the
ObjectField verifier to perform a manual walk of the ClassOp to find the target
field. In the future, we should switch to InnerSymbols or invent something
totally new for these operations.