[OM][LowerClasses] Insert lowered ClassOps into the FIRRTL circuit
OM path operations use a HierPathOp to identify their target.  To facilitate
this, it makes sense to insert OM classes under the FIRRTL circuit, as
HierPathOp should be under the same symbol table as the path operation for
lookup to work correctly. If they were not under the same symbol table, it
would be possible to craft a nested symbol reference, but that would just
complicate the LowerToHw pass.

This change removes the restriction that OM ClassOps have MLIRModuleOp as a
parent, changes LowerClasses from a MLIRModule pass to a FIRRTL CircuitOp pass,
and updates some logic to ensure that we are inserting the class operations in
to the correct location.

As a quick example of how this will look with path operations:

```mlir
module { // SymbolTable
  firrtl.circuit "Foo" { // SymbolTable
    hw.hierpath @Path [@Foo::@Wire]
    firrtl.module @Foo { // Not a SymbolTable
        %wire = firrtl.wire sym @wire : !firrtl.uint<1>
    }
    om.class @Class() { // Not a SymbolTable
      %0 = firrtl.path @Path
    }
  }
}
```