[HW] HierPathOp: support path to modules
This changes the HierPathVerifier to support single-element paths which point
to modules.  HierPathOps already support single-length paths to InnerSymbols,
and we just needed to copy the logic to skip checking that the module matches
the expected module name when the path is a single element.  Missing support is
most likely a bug in the initial implementation, as the error message printed a
null attribute.