[FIRRTL] Refactor BlackBoxReader to get/set, NFC
Refactor the internals of the BlackBoxReader pass to break the
setOutputFile member function into two functions: (1) getOutputFile to
determine the output file and (2) setOutputFile to set the output file.
This is done in preparation to enable better checking that one extmodule
implemented as a black box is not clobbering another extmodule with the
same defname, but a different file.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>