[FIRRTL] Update LowerClasses to handle extmodule and instances.
This adds lowering for extmodule similar to extclasses and module: an
external class declaration is created with the property input and
output ports. This also adds support for converting instances of
modules or extmodules that have been lowered, and keeping the instance
graph up to date.

A small addition is also included to always create a class for public
modules and extmodules, which is a further refinement of the as-yet
undocumented ABI.