[FIRRTL] Update LowerClasses to handle property wires.
Property wires are added in some scenarios for convenience. This would
be a good canonicalization pattern to have, but canonicalization
shouldn't be needed for correctness. So, this pass explicitly handles
wires of properties, by simply finding the single assignment to the
wire, using that value instead of the wire, and removing the
assignment.

Closes https://github.com/llvm/circt/issues/6051.