[FIRRTL] Handle duplicate BlackBoxAnnos
Fix a bug in FIRRTL's BlackBoxReader pass that could manifest if multiple
blackbox inline or blackbox path annotations wanted to write to the same
file.  Previously, the first annotation would win and all remaining
annotations would be skipped.  However, each of these annotations may have
different "output directory priority" and the first one that is found may
write to the wrong directory.

Concretely, these annotations can cause writes to any of the main output
directory (where "DUT" files go), the verification extracted directory,
the Grand Central directory, or the testbench directory.  The main output
directory is the "highest" priority and the testbench directory is the
"lowest" priority.  If a blackbox is in both, then it should be written to
the one with higher priority.

Change the logic of BlackBoxReader to collect all annotations that
indicate that they want to write to a file with the same name and then
determine which of them gets to based on the priorities defined above.

Fixes #6035.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>