[FIRRTL] Update LowerClasses to handle property wires. (#6067)
Property wires are added in some scenarios for convenience. This would
be a good canonicalization pattern to have, but canonicalization
shouldn't be needed for correctness, and this isn't always possible in
non-graph regions. So, this pass explicitly handles wires of
properties when they are known to be in graph regions, by simply
finding the single assignment to the wire, using that value instead of
the wire, and removing the assignment.

Closes #6051.