[Ibis] Add `--ibis-argify-blocks` pass (#6084)
The `ibis.block` operation defines a block wherein a group of operations are expected to be statically scheduleable. The operation is not isolated from above to facilitate ease of construction. However, once a program has been constructed and lowered to a sufficient level, the user may run `--ibis-argify-blocks` to effectively isolate the block from above, by converting SSA values referenced through dominanes into arguments of the block

The block may contain additional attributes to specify constraints on the block further down the compilation pipeline.

This operation subsumes the `ibis.schedule` in that `ibis.schedule` has the same semantics as if we attach some scheduling constraint attribute to the block.