[Pipeline] Fix issue in ExplicitRegs
The order of pass and reg operands was incorrect in between the order that block arguments were added to a stage, and the order that said block arguments were used to replace backedges within a block. Up until now, we were just lucky that our existing tests didn't actually infer any interleaving of pass and reg values flowing through the pipeline.