[FIRRTL] Handle zero-width values escaping to foreign ops
At the moment the FIRRTL-to-HW lowering removes zero-width values
everywhere. This works as long as all ops involved are lowered by
LowerToHW and have the zero-width value removed. If such a value is used
by a foreign op however, it escapes LowerToHW's ability to be properly
removed. In this case, instead of crashing, create an
`hw.constant 0 : i0` value to pass into the foreign op.