[ExportVerilog] Remove useless assert from NameCollector
The `NameCollector` is only used to compute declaration word and type
string length such that we can format declarations nicely. Currently it
also asserts a weird invariant that `PrepareForEmission` must uphold,
but then it never actually relies on that invariant. And for some reason
the `NameCollector` also collects names in else branches of `IfOp`s, but
not in then branches.

Push `NameCollector` more into the best-effort cosmetics direction it's
already going by removing that assert and not visiting else branches.
Either we visit both branches of an if, or none at all. And since if
blocks introduce a separate scope with a separate place for decls, the
visition is not necessary. (It would just potentially add whitespace to
the if's surrounding scope's decl stack, to account for decls that are
actually emitted within the if scope.)

This also uncovered one instance of `maxDeclNameWidth - x` that was left
unguarded by `if (x < maxDeclNameWidth)`. Fix this.

This change doesn't affect the output, but makes ExportVerilog less
brittle by not enforcing invariants that it actually doesn't care about.