[FirRegLowering] Implement refined check for mux-to-if conversion.
This consults the register's fanout, and only pulls muxes into if/else
structure if they are on the register's fanout. We want the if/else
structure to infer enable signals, but we don't want the other muxes
to be pulled into if/else structure if they aren't related to the
register's enable.

This is implemented as a forward dataflow slice from the register for
efficiency reasons. We can take one slice per register before we start
building the tree, and consult it for each mux we encounter as we move
back from the register. The upstream forward dataflow slice is
recursive, and will cause stack overflows in practice, so an iterative
version is vendored into this pass.