[HandshakeToDC] Turn conversion into library-style
Factors out the conversion s.t. it can be run on arbitrary operations. To facilitate this, the conversion will accept builders for additional patterns to apply during the converison process. Having this, users may specify patterns for specific ops (to avoid applying the default unit-rate actor pattern to said ops).