[ExportVerilog] Bound type size considered for decl alignment (#6171)
We currently blindly align the types and names of declarations, which
causes excessively large types like structs to ruin the alignment of
declarations. Instead, introduce an upper bound above which a type does
not contribute towards the `maxTypeWidth` count. This means that
excessively large types are just emitted without alignment, without
other declarations being forced to follow the same alignment. This
change is purely cosmetic.

Also tweak the code that emits the spacing around declarations to work
by specifying a target column up to which spaces should be filled, and
then filling in those spaces, or a minimum of 1. This makes it harder
to get the alignment off by 1 just because some type or decl word was
empty and didn't need a trailing space.