[Comb] Disable canonicalization across blocks
This is probably the most conservative implementation of this, but should suffice for the current requirements to this change.
For any comb operation that has a canonicalizer, guard canonicalization on whether any operands are define outside of the current block.
This still allows for (constant) folding, which (to me) still should be a safe thing to do across blocks - if that was not the case, then that would be a strange abstraction (i'd expect an op to be `IsolatedFromAbove` if it wants to prevent constant folding).