[OM] Evaluator: Support graph regions
This commit adds support for graph regions for evaluator.

`ReferenceValue`, a new subclass of `EvaluatorValue`, is added to behave
as pointers. `ReferenceValue` can be used as alias to different values and
is created for `class.object.field` operation because `class.object.field`
can access fields across class hierarchies and the fields might not be
initilized yet. `RefenceValue` is not exposed to outside of evaluator implemenation.
`EvaluatorValue::finalize` shrinks intermidiate `RefenceValue` in the evaluator
value.

Evaluation algorithm is changed to worklist-based iteration. `instantiae`
method first traverses the whole IR including sub-class, and create partially evaluaed
values for all values and add these values to the worklist. After that
we evaluate values until there is no partially evaluaed value.