[OM] Add utility for parsing target strings
This adds an OM utility for parsing FIRRTL style target strings, with the
intention of using them in the frozen path MLIR assembly format.  The existing
APIs contained in FIRRTL were not suitable since these new strings do not
include the circuit name, and we need some restrictions for parsing base paths,
which can not target components. The new parser is a bit more resilient to
badly formed path strings. In the future it may be possible to reunite these
two bodies of code.