[Seq] Add optional power-on value to `compreg` ops
In doing so, also opt to move to a keyword style assembly format, rather than a flat list of operands. This is needed to resolve the now three different, optional, operands that can occur in the assembly format (reset operands, clock-enable, and the power-on operand).
To avoid making things too complicated, a keyword is required for either group, e.g.
> `seq.compreg %in %clk reset %rst, %rstValue ce %ce powerOn %po`

+ some related cleanups in places where the default generated builder was being used (without good reason).