[FIRRTL][FIRParser] Prefer RWProbe op as much as possible. (#5835)
Keep forceable for now-- it's needed for uninferred resets presently,
and has slightly better (if not ideal) behavior when attempting
to rwprobe a non-passive aggregate.

Accordingly the spec test `force_nonpassive.fir` now fails.

We can't rwprobe() an aggregate that /must/ be lowered.  Forceable
tries to allow this, but when using symbols this is just rejected.

Force of non-passive will soon be unsupported and spec updated to make illegal.