[Namespace] Return an empty string for empty string.
Previously Namespace::newName could return empty string for empty string.
This is a problematic behaivor since primary use cases of Namespace is
symbol generation but empty string cannot be used a symbol.

For example HWExportModuleHierarchy crashed with an instance with an
empty name.

This commit fixes the issue by pre-registering empty string in a constructor of
Namespace.