[HW] Fix ssa name coupling on module ports.
MLIR is free to rename ssa values on module ports.  It will compute that name from the port name using its legalization algorithm.  This means single-round-trips are not stable, but subsequent ones are.  This is consistent with the behavior of other ops which have value name hints (e.g. you can have your own value at parse time, but it will be computed from scratch in the output).