[Debug] Include Verilog instance names in HGLDD output
If an instance's Verilog name differs from the debug info instance name,
add an additional `hdl_obj_name` field to the HGLDD output. Currently,
the only way to trigger this is to have an instance name be a Verilog
keyword, such that name legalization appens a `_N` suffix. In the
future, with inlining and artificial levels of hierarchy, there are
likely going to be more ways of how a DI instance name can differ from
the Verilog hierarchy.