[ESI] The dialect conversion framework can't be used in bundle lowering
The dialect conversion framework tries to legalize everything touched by
a pattern, not just things created by one in order to consider the
original operation legalized. Additionally, it won't run the same
pattern twice. So when one has an illegal operation which is being
eliminated and replacing a value which happens to feed into another
illegal operation which itself gets legalized by the same pattern, the
pattern doesn't get run on the illegal consumer and the original op
isn't considered legalized.