[CI] Update Python wheel action for MacOS.
The macos-latest tag has been migrated to macos-12, and the version of
cibuildwheel we use doesn't run with the version of Python included in
macos-12 (Python 3.12). Note that this has no bearing on the versions
of Python we build the wheels for; this is the cibuildwheel Python
code that is running on the runner's default Python to set up the
build for the version of Python the wheel is targeting.

To amend this, I am locking to MacOS tag to a specific version, and
updating the cibuildhweel version.

Fixes https://github.com/llvm/circt/issues/6385.