[ESI][Runtime] Read a manifest and build the design hierarchy (#6384)
Start of manifest-based runtime API construction. The manifest is parsed from a string, which could be read from a file or the live accelerator. We then construct the design hierarchy given a connection to the accelerator. A connection is not necessary for this PR, but it will be necessary for the next PR which will instantiate connections to the services.

As of now, the Python interface is just the raw pybound11 classes, though this'll get better over time. Also, the pyi file was generated by pybind11-stubgen but I am committing it since that's what MLIR does. For some reason they've had to do some manual modifications, so I assume we'll have to as well in the future.