[FIRRTL] LowerClasses: rebase paths through module instances
In an effort to support multiple instantiation of modules, we need to overhaul
the ABI for paths.  Before, it was accidentally assumed that only public or
top-level modules could be multiply instantiated, and base-paths would be
rebased at those boundaries.  Now, when lowering modules to classes (but not
when classes are lowered to classes), the basepath is always rebased.  This
lets us compose paths across multiple module instances.